epsg_number to epsg_string FeatureType to FeatureTypeSet ifel to if, added returns
