+++ .travis.yml
@@ -24,13 +24,15 @@ install:
 - pip install pandas==0.23.4  # This is temporal solution
 
 script:
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint core/eolearn/core/*.py ; fi
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint coregistration/eolearn/coregistration/*.py ; fi
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint features/eolearn/features/*.py ; fi
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint geometry/eolearn/geometry/*.py ; fi
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint io/eolearn/io/*.py ; fi
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint mask/eolearn/mask/*.py ; fi
-- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then pylint ml_tools/eolearn/ml_tools/*.py ; fi
+- if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then
+  pylint core/eolearn/core/*.py
+  pylint coregistration/eolearn/coregistration/*.py
+  pylint features/eolearn/features/*.py
+  pylint geometry/eolearn/geometry/*.py
+  pylint io/eolearn/io/*.py
+  pylint mask/eolearn/mask/*.py
+  pylint ml_tools/eolearn/ml_tools/*.py
+fi
 - radon cc core/eolearn/core/*.py -a -nb
 - radon cc coregistration/eolearn/coregistration/*.py -a -nb
 - radon cc features/eolearn/features/*.py -a -nb
